nested routes for Manufacturers #1154
