+++ src/App.tsx
@@ -18,14 +18,16 @@ import {
 import { MicroFrontendId } from './app.types';
 import Catalogue from './catalogue/catalogue.component';
 import CatalogueItemsLandingPage from './catalogue/items/catalogueItemsLandingPage.component';
+import ErrorPage from './common/errorPage.component';
 import ConfigProvider from './configProvider.component';
 import handleIMS_APIError from './handleIMS_APIError';
 import { HomePage } from './homePage/homePage.component';
 import IMSThemeProvider from './imsThemeProvider.component';
 import Items from './items/items.component';
 import ItemsLandingPage from './items/itemsLandingPage.component';
-import ManufacturerComponent from './manufacturer/manufacturer.component';
+import Manufacturer from './manufacturer/manufacturer.component';
 import ManufacturerLandingPage from './manufacturer/manufacturerLandingPage.component';
+import ManufacturerTable from './manufacturer/manufacturersTable.component';
 import Preloader from './preloader/preloader.component';
 import retryIMS_APIErrors from './retryIMS_APIErrors';
 import {
@@ -88,10 +90,22 @@ const router = createBrowserRouter([
         Component: ItemsLandingPage,
       },
       { path: paths.systems, Component: Systems },
-      { path: paths.manufacturers, Component: ManufacturerComponent },
       {
-        path: paths.manufacturer,
-        Component: ManufacturerLandingPage,
+        path: paths.manufacturers,
+        Component: Manufacturer,
+        children: [
+          { path: '', Component: ManufacturerTable },
+          { path: paths.manufacturer, Component: ManufacturerLandingPage },
+          {
+            path: '*',
+            Component: () => (
+              <ErrorPage
+                boldErrorText="Invalid Manufacturer Route"
+                errorText="The manufacturer route you are trying to access doesn't exist. Please click the Home button to navigate back to the Manufacturer Home page."
+              />
+            ),
+          },
+        ],
       },
     ],
   },
